refactor: move sub parser into dedicated modules
Previously, sub-module parsers were located in src/debsbom/cli.py.
To simplify maintenance, each command's parser with command executor
have been relocated to their own dedicated file within
src/debsbom/commands.
To avoid circular import issues, SbomInput and PkgsStreamInput were
moved to src/debsbom/commands/input.py, and progress_cb was relocated to
src/debsbom/util/progress.py.
