+++ src/debsbom/download/download.py
@@ -10,11 +10,15 @@ from hmac import compare_digest
 import logging
 import shutil
 from pathlib import Path
-import requests
 
 from ..dpkg import package
 from ..dpkg.package import ChecksumAlgo
 from ..snapshot.client import RemoteFile
+
+try:
+    import requests
+except ModuleNotFoundError:
+    pass
 
 
 logger = logging.getLogger(__name__)
